[Update] 오늘 날짜 반환하는 함수 MyGlobal에서 Config 파일로 이동
