+++ MediaFIT-master/Exercise_Helper/app/src/main/java/com/gauravk/bubblebarsample/cfg/MyGlobal.java
@@ -123,38 +123,4 @@ public class MyGlobal {
     public void setPOSE_SAMPLE_FILE(String str){this.POSE_SAMPLE_FILE = str;}
     public void setFinish(boolean fin){this.finish = fin;}
 
-
-    public static int index = -1;
-    public static String today_hangle(){
-        Calendar cal = Calendar.getInstance();
-        index = cal.get(Calendar.DAY_OF_WEEK);
-        String WhatWeek;
-        switch (index){
-            case 1:
-                WhatWeek = "일";
-                break;
-            case 2:
-                WhatWeek = "월";
-                break;
-            case 3:
-                WhatWeek = "화";
-                break;
-            case 4:
-                WhatWeek = "수";
-                break;
-            case 5:
-                WhatWeek = "목";
-                break;
-            case 6:
-                WhatWeek = "금";
-                break;
-            case 7:
-                WhatWeek = "토";
-                break;
-            default:
-                WhatWeek = "이상하다";
-                break;
-        }
-        return WhatWeek;
-    }
 }
