+++ MediaFIT-master/Exercise_Helper/app/src/main/res/layout/fragment_routine_list.xml
@@ -33,9 +33,8 @@
     </LinearLayout>
 
 
-
     <include
-        layout="@layout/routine_button"/>
+        layout="@layout/routine_button" />
 
     <include
         layout="@layout/fragment_routine_content" />
